src/backend/requirements.txt
fastapi
uvicorn
motor
pydantic
bcrypt
python-jose
passlib[bcrypt]
python-dotenv